Add .editorconfig, add some basic config options

+++ StatusTimers/Extensions/CharacterExtensions.cs
@@ -4,14 +4,14 @@ using FFXIVClientStructs.FFXIV.Client.Game.Character;
 
 namespace StatusTimers.Extensions;
 
-public static class CharacterExtensions
-{
-    public static unsafe bool IsHostile(this ICharacter character)
-    {
+public static class CharacterExtensions {
+    public static unsafe bool IsHostile(this ICharacter character) {
         Character* chara = (Character*)character.Address;
 
         return character != null
-               && ((character.SubKind == (byte)BattleNpcSubKind.Enemy || (int)character.SubKind == (byte)BattleNpcSubKind.BattleNpcPart)
-                   && chara->CharacterData.Battalion > 0); // Since its not super clear, CharacterData.Battalion used for determining friend/enemy state
+               && (character.SubKind == (byte)BattleNpcSubKind.Enemy ||
+                   character.SubKind == (byte)BattleNpcSubKind.BattleNpcPart)
+               && chara->CharacterData.Battalion >
+               0; // Since its not super clear, CharacterData.Battalion used for determining friend/enemy state
     }
 }
